tests/requirements.txt
pytest
requests
pyyaml
pytest
pytest-sugar
